+++ src/main/java/defencer/controller/entity/ProjectController.java
@@ -3,6 +3,7 @@ package defencer.controller.entity;
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXComboBox;
 import com.jfoenix.controls.JFXDatePicker;
+import defencer.controller.AskFormController;
 import defencer.controller.PremierLeagueController;
 import defencer.controller.update.UpdateProjectController;
 import defencer.data.ControllersDataFactory;
@@ -39,6 +40,7 @@ import java.util.ResourceBundle;
  * @author Igor Gnes on 4/13/17.
  */
 public class ProjectController implements Initializable {
+
 
     @FXML
     private TableView<Project> table;
@@ -177,7 +179,6 @@ public class ProjectController implements Initializable {
 
     /**
      * Open new page to add one more project.
-     *
      * {@link SneakyThrows} here because i am totally sure that path to fxml is correct.
      */
     @SneakyThrows
@@ -196,7 +197,6 @@ public class ProjectController implements Initializable {
         stage.initOwner(window);
         stage.show();
 
-        stage.setOnHiding(e -> loadProjects());
     }
 
     /**
@@ -273,9 +273,36 @@ public class ProjectController implements Initializable {
             return;
         }
         try {
-            ServiceFactory.getProjectService().deleteEntity(project);
-            observableProjects.clear();
-            loadProjects();
+
+            final FXMLLoader fxmlLoader = new FXMLLoader();
+            fxmlLoader.setLocation(getClass().getResource("/askForm.fxml"));
+            final Parent parent = fxmlLoader.load();
+            final Stage stage = new Stage();
+            stage.setTitle("Patriot Defence");
+            Scene value = new Scene(parent);
+            value.getStylesheets().add("css/main.css");
+            stage.setScene(value);
+            stage.initModality(Modality.WINDOW_MODAL);
+            Window window = btnDelete.getScene().getWindow();
+            stage.initOwner(window);
+            stage.show();
+
+            stage.setOnHiding(event -> {
+
+                System.out.println(ControllersDataFactory.getLink().get(AskFormController.class, "isDelete"));
+
+                if ((boolean) ControllersDataFactory.getLink().get(AskFormController.class, "isDelete")) {
+
+                    try {
+                        ServiceFactory.getProjectService().deleteEntity(project);
+                    } catch (Exception e) {
+                        NotificationUtil.errorAlert("Error", "Can't delete", NotificationUtil.SHORT);
+                    }
+                    observableProjects.clear();
+                    loadProjects();
+                }
+            });
+
         } catch (Exception e) {
             NotificationUtil.errorAlert("Error", "Can't delete", NotificationUtil.SHORT);
         }
